feat(products): add variant SKU routing and price improvements
- Auto-generate SKUs for variants from product name + color/size
- Add variant-level URL routing (/producto/{slug}/{variantSku?})
- Replace browser history state on variant selection
- Apply ScaledPrice cast to Product and ProductVariant models
- Update formatPrice to accept float instead of cents
- Add price_cost and price_provider fields to product form
- Show existing variant combinations in generator modal
- Add out-of-stock state to add-to-cart button
- Expand CSS design tokens (type scale, color palette, font weights)
- Rename docker containers from laravel-* to project-specific names
- Add collapsible sidebar to Filament backend panel

+++ app/Services/CartService.php
@@ -103,7 +103,7 @@ class CartService
         session()->forget('cart_coupon');
     }
 
-    public function getSubtotal(): int
+    public function getSubtotal(): float
     {
         $cart = session('cart', []);
         if (empty($cart)) {
@@ -168,7 +168,7 @@ class CartService
         return $productId.'-'.($variantId ?? 'null');
     }
 
-    private function getItemPrice(?Product $product, ?ProductVariant $variant): int
+    private function getItemPrice(?Product $product, ?ProductVariant $variant): float
     {
         if (! $product) {
             return 0;
